// src/types.d.ts
export interface Input {
  type: HTMLInputTypeAttribute | 'select';
  name: string;
  label?: string;
  autocomplete?: string;
  placeholder?: string;
  /** For type="select" */
  options?: Array<{ value: string; label: string }>;
  required?: boolean;
}

export interface Textarea {
  label?: string;
  name?: string;
  placeholder?: string;
  rows?: number;
  /** Defaults to true for inquiry forms */
  required?: boolean;
}

export interface Disclaimer {
  label?: string;
}